+++ server/lib/tuist/marketing/pages/page_parser.ex
@@ -5,15 +5,19 @@ defmodule Tuist.Marketing.Pages.PageParser do
   like the homepage, about page, etc.
   """
 
+  alias Tuist.Marketing.MDExConverter
+
   def parse(path, contents) do
     [frontmatter_string, body] =
-      contents |> String.replace(~r/^---\n/, "") |> String.split(["\n---\n"])
+      contents |> String.replace(~r/^---\n/, "") |> String.split(["\n---\n"], parts: 2)
 
     frontmatter =
       frontmatter_string
       |> YamlElixir.read_from_string!()
       |> Map.put("slug", "/#{path |> Path.basename() |> String.replace(".md", "")}")
 
-    {frontmatter, body}
+    {body_html, _body_template} = MDExConverter.compile_markdown(body, path, false)
+
+    {frontmatter, body_html}
   end
 end